+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 **Version: 0.0.76 (BETA)**
 
 * Mark Container packet handled so it does not cause a print statement.
+* New Models for some blocks
+* Multi-Tool Assist render now shows the placed state of block
 
 -------------------------------------------------------------------
 **Version: 0.0.75 (BETA)**